tests/requirements.txt
# Test dependencies for PyLLMs
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-timeout>=2.1.0
pytest-mock>=3.10.0
pytest-cov>=4.0.0
pytest-xdist>=3.0.0  # For parallel test execution 
python-dotenv>=1.0.0  # For automatic .env loading